hr/requirements.txt
Django==5.0.7
pillow
djangorestframework
requests
django-cors-headers

